app/requirements.txt
flask
flask-login
flask-restful
pyserial